# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Radiation Tracking Backend",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/history/timeseries": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Per-window CPM time-series (count-weighted) for the trend chart",
        "operationId": "get_timeseries_history_timeseries_get",
        "parameters": [
          {
            "name": "min_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lat"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lat"
            }
          },
          {
            "name": "min_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lon"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lon"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeseriesPoint"
                  },
                  "title": "Response Get Timeseries History Timeseries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/history/hotspots": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Top geohash cells by peak CPM over the range",
        "operationId": "get_hotspots_history_hotspots_get",
        "parameters": [
          {
            "name": "min_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lat"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lat"
            }
          },
          {
            "name": "min_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lon"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lon"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotspot"
                  },
                  "title": "Response Get Hotspots History Hotspots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hotspot": {
        "properties": {
          "geohash": {
            "type": "string",
            "title": "Geohash"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "peak_cpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Cpm"
          },
          "avg_cpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cpm"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "default": 0
          },
          "windows": {
            "type": "integer",
            "title": "Windows",
            "default": 0
          },
          "anomaly_windows": {
            "type": "integer",
            "title": "Anomaly Windows",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "geohash"
        ],
        "title": "Hotspot"
      },
...
      "TimeseriesPoint": {
        "properties": {
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "avg_cpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cpm"
          },
          "max_cpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cpm"
          },
          "danger": {
            "type": "integer",
            "title": "Danger",
            "default": 0
          },
          "anomaly_cells": {
            "type": "integer",
            "title": "Anomaly Cells",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "window_start",
          "total_count"
        ],
        "title": "TimeseriesPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}